+++ NEWS.md
@@ -2,7 +2,9 @@
 
 # development version
 
-- Add Hyland 2009 midazolam N-glucuronidation ([doi:10.1111/j.1365-2125.2009.03386.x](https://doi.org/10.1111/j.1365-2125.2009.03386.x)) -- in vitro, pooled human liver microsomes and recombinant UGT1A4.
+- Add Benson 2010 reboxetine hNET target-binding kinetics
+  ([doi:10.1111/j.1476-5381.2010.00719.x](https://doi.org/10.1111/j.1476-5381.2010.00719.x))
+  -- in vitro, HEK-293 membranes expressing human noradrenaline transporter.
 
 - Enrich `Lahu_2010_roflumilast` provenance from the FDA NDA 22-522 Clinical
   Pharmacology review, which reprints the sponsor's population PK study report